feat : enum 따로 빼기
