+++ src/main/java/com/example/ootd/batch/listener/WeatherAlertListener.java
@@ -38,24 +38,6 @@ public class WeatherAlertListener extends StepListenerSupport<WeatherBatchData, 
   @Autowired
   private UserRepository userRepository;
 
-  // 날씨 알림 유형 정의
-  public enum WeatherAlertType {
-    HEAT_WAVE("폭염 경보"),
-    HEAVY_RAIN("호우 경보"),
-    SNOW_WARNING("대설 경보"),
-    STRONG_WIND("강풍 경보");
-
-    private final String message;
-
-    WeatherAlertType(String message) {
-      this.message = message;
-    }
-
-    public String getMessage() {
-      return message;
-    }
-  }
-
   @Override
   public void afterRead(WeatherBatchData item) {
     log.debug("Successfully read data for region: {}", item.getRegionName());
